+++ src/web/context_processors.py
@@ -1,17 +1,21 @@
 from django.conf import settings
 from django.contrib.auth.views import LoginView
+from django.http import HttpRequest
 from django.utils.translation import get_language
 from django_hosts import reverse, reverse_host
 
 
-def common_context_variables(request):
+def common_context_variables(request: HttpRequest) -> dict:
     return {
+        # Language-related stuff
         'DEFAULT_LANGUAGE_CODE': settings.LANGUAGE_CODE,
         'CURRENT_LANGUAGE_CODE': get_language(),
+        # Misc. stuff
+        'IS_DEV_ENV': settings.IS_DEV_ENV,
     }
 
 
-def login(request):
+def login(request: HttpRequest) -> dict:
     # The current `next` parameter
     current_redirect_url = LoginView(request=request).get_redirect_url()